requirements.txt
fastapi
uvicorn
opentelemetry-instrumentation
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-asyncio==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-sqlite3==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-instrumentation-click==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-grpc==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-starlette==0.52b1
opentelemetry-instrumentation-tortoiseorm==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
